[JENKINS-43845] - Rework POM and move contents to a new package
