+++ pom.xml
@@ -1,76 +1,61 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
 
-    <modelVersion>4.0.0</modelVersion>
+  <modelVersion>4.0.0</modelVersion>
 
-    <parent>
-        <groupId>org.jenkins-ci</groupId>
-        <artifactId>jenkins</artifactId>
-        <version>1.29</version>
-    </parent>
+  <parent>
+    <groupId>org.jenkins-ci.plugins</groupId>
+    <artifactId>plugin</artifactId>
+    <version>2.26</version>
+  </parent>
 
-    <groupId>org.jenkins-ci.lib</groupId>
-    <artifactId>envinject-lib</artifactId>
-    <version>1.25-SNAPSHOT</version>
-    <packaging>jar</packaging>
-    <name>Jenkins EnvInject Library</name>
-    <url>https://wiki.jenkins-ci.org/display/JENKINS/EnvInject+Plugin</url>
+  <groupId>org.jenkins-ci.lib</groupId>
+  <artifactId>envinject-api</artifactId>
+  <version>1.25-SNAPSHOT</version>
+  <packaging>jar</packaging>
+  <name>EnvInject API Plugin</name>
+  <description>Stores shared logic for Environment Injection logic</description>
+  <url>https://wiki.jenkins-ci.org/display/JENKINS/EnvInject+API+Plugin</url>
 
-    <licenses>
-        <license>
-            <name>MIT license</name>
-            <comments>All source code is under the MIT license.</comments>
-        </license>
-    </licenses>
+  <licenses>
+    <license>
+      <name>MIT license</name>
+      <comments>All source code is under the MIT license.</comments>
+    </license>
+  </licenses>
 
-    <developers>
-        <developer>
-            <id>gbois</id>
-            <name>Gregory Boissinot</name>
-            <timezone>+1</timezone>
-            <roles>
-                <role>Project Owner</role>
-                <role>Project Lead Developer</role>
-            </roles>
-        </developer>
-    </developers>
+  <developers>
+    <developer>
+      <id>oleg_nenashev</id>
+      <name>Oleg Nenashev</name>
+      <timezone>+1</timezone>
+    </developer>
+  </developers>
 
-    <properties>
-        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven.compiler.source>1.6</maven.compiler.source>
-        <maven.compiler.target>1.6</maven.compiler.target>
-        <jenkins.core.version>1.532</jenkins.core.version>
-    </properties>
+  <properties>
+    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+    <java.level>6</java.level>
+    <jenkins.version>1.532</jenkins.version>
+  </properties>
 
-    <scm>
-        <connection>scm:git:git://github.com/jenkinsci/envinject-lib.git</connection>
-        <developerConnection>scm:git:git@github.com:jenkinsci/envinject-lib.git</developerConnection>
-      <tag>HEAD</tag>
+  <scm>
+    <connection>scm:git:git://github.com/jenkinsci/envinject-lib.git</connection>
+    <developerConnection>scm:git:git@github.com:jenkinsci/envinject-lib.git</developerConnection>
+    <tag>HEAD</tag>
   </scm>
 
-    <dependencies>
+  <repositories>
+    <repository>
+      <id>repo.jenkins-ci.org</id>
+      <name>Jenkins Repository</name>
+      <url>https://repo.jenkins-ci.org/public/</url>
+    </repository>
+  </repositories>
 
-        <dependency>
-            <groupId>org.jenkins-ci.main</groupId>
-            <artifactId>jenkins-core</artifactId>
-            <version>${jenkins.core.version}</version>
-            <scope>provided</scope>
-        </dependency>
-
-    </dependencies>
-
-    <repositories>
-        <repository>
-            <id>repo.jenkins-ci.org</id>
-            <name>Jenkins Repository</name>
-            <url>http://repo.jenkins-ci.org/public/</url>
-        </repository>
-    </repositories>
-
-    <pluginRepositories>
-        <pluginRepository>
-            <id>repo.jenkins-ci.org</id>
-            <url>http://repo.jenkins-ci.org/public/</url>
-        </pluginRepository>
-    </pluginRepositories>
+  <pluginRepositories>
+    <pluginRepository>
+      <id>repo.jenkins-ci.org</id>
+      <url>https://repo.jenkins-ci.org/public/</url>
+    </pluginRepository>
+  </pluginRepositories>
 
 </project>
