+++ src/main/java/org/jenkinsci/plugins/envinjectapi/EnvInjectLogger.java
@@ -1,4 +1,4 @@
-package org.jenkinsci.lib.envinject;
+package org.jenkinsci.plugins.envinjectapi;
 
 import hudson.model.TaskListener;
